requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
